# docs/contracts/openapi-area-jugador.yaml
openapi: 3.1.0
info:
  title: ReservaYa API — Contrato de Área Jugador y Vitrina Pública
  description: >
    Especificación OpenAPI de los contratos compartidos entre la landing pública
    (apps/web), el panel de usuario (apps/web) y
    el backend (.NET 10 ReservaFacil.Api) para el área del jugador bajo
    arquitectura desacoplada.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Backend local .NET
  - url: http://localhost:3000
    description: Proxy Next.js local
  - url: https://reservaya-api.onrender.com
    description: Backend Render producción
paths:
...
  /api/auth/me:
    get:
      summary: Obtener perfil de la sesión actual autenticada
      responses:
        '200':
          description: Sesión válida
          content:
            application/json:
              schema:
                type: object
                required: [usuario]
                properties:
                  usuario:
                    $ref: '#/components/schemas/UsuarioSesion'
        '401':
          description: Sin sesión activa
  /api/auth/login:
    post:
      summary: Iniciar sesión con email y contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login exitoso, cookie Set-Cookie emitida
          content:
            application/json:
              schema:
                type: object
                required: [ok, usuario]
                properties:
                  ok:
                    type: boolean
                  usuario:
                    $ref: '#/components/schemas/UsuarioSesion'
        '401':
          description: Credenciales incorrectas
        '429':
          description: Rate limit excedido
  /api/auth/register:
    post:
      summary: Registrar nuevo jugador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, email, password, fechaNacimiento, username]
              properties:
                nombre:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                fechaNacimiento:
                  type: string
                  format: date
                username:
                  type: string
      responses:
        '200':
          description: Registro exitoso
          content:
            application/json:
              schema:
                type: object
                required: [ok, usuario]
                properties:
                  ok:
                    type: boolean
                  usuario:
                    $ref: '#/components/schemas/UsuarioSesion'
...
components:
  schemas:
...
    UsuarioSesion:
      type: object
      required: [id, email, nombre, rol]
      properties:
        id:
          type: string
        email:
          type: string
        nombre:
          type: string
        username:
          type: string
          nullable: true
        fechaNacimiento:
          type: string
          nullable: true
        telefono:
          type: string
          nullable: true
        fotoUrl:
          type: string
          nullable: true
        rol:
          type: string
          enum: [USUARIO, ADMIN, SUPERADMIN, TECNICO]
...